move data injection to separate file & fix source tar
